Migrate regressions from 2000-2500
